+++ apps/backend/src/modules/user-follows/libs/exceptions/user-follows.exception.ts
@@ -1,0 +1,5 @@
+import { HTTPError } from "~/libs/modules/http/http.js";
+
+class UserFollowsError extends HTTPError {}
+
+export { UserFollowsError };