+++ apps/frontend/src/index.tsx
@@ -16,7 +16,9 @@ import { Auth } from "~/pages/auth/auth.jsx";
 import { NotFound } from "~/pages/not-found/not-found.jsx";
 
 import { Dashboard } from "./pages/dashboard/dashboard.js";
+import { Explore } from "./pages/explore/explore.js";
 import { Landing } from "./pages/landing/landing.jsx";
+import { Profile } from "./pages/profile/profile.js";
 
 pwa.register();
 
@@ -50,6 +52,18 @@ createRoot(document.querySelector("#root") as HTMLElement).render(
 									</ProtectedRoute>
 								),
 								path: AppRoute.ADMINISTRATIVE_DASHBOARD,
+							},
+							{
+								element: <Explore />,
+								path: AppRoute.EXPLORE,
+							},
+							{
+								element: (
+									<ProtectedRoute>
+										<Profile />
+									</ProtectedRoute>
+								),
+								path: AppRoute.PROFILE,
 							},
 						],
 						element: <App />,