+++ apps/frontend/src/libs/enums/app-route.enum.ts
@@ -1,7 +1,12 @@
 const AppRoute = {
 	ADMINISTRATIVE_DASHBOARD: "/app/administrative-dashboard",
 	APP: "/app",
+	CATEGORIES: "/app/categories",
+	CONTENT: "/app/content",
+	EXPLORE: "/app/explore",
+	PROFILE: "/app/profile",
 	ROOT: "/",
+	ROUTES: "/app/routes",
 	SIGN_IN: "/app/sign-in",
 	SIGN_UP: "/app/sign-up",
 } as const;