+++ apps/frontend/src/libs/enums/event-listener-type.enum.ts
@@ -1,0 +1,6 @@
+const EventListenerType = {
+	POPSTATE: "popstate",
+	WHEEL: "wheel",
+} as const;
+
+export { EventListenerType };