+++ apps/frontend/src/libs/modules/map-client/libs/constants/map-controls-position.constant.ts
@@ -1,0 +1,9 @@
+import { type ControlPosition } from "../types/types.js";
+
+const MAP_CONTROLS_POSITION: Record<string, ControlPosition> = {
+	GEOLOCATE: "bottom-right",
+	NAVIGATION: "top-right",
+	SCALE: "bottom-left",
+};
+
+export { MAP_CONTROLS_POSITION };