+++ apps/frontend/src/libs/modules/map-client/libs/constants/zoom-level.constant.ts
@@ -1,0 +1,3 @@
+const ZOOM_LEVEL = 12;
+
+export { ZOOM_LEVEL };