+++ apps/frontend/src/libs/modules/map-client/libs/enums/enums.ts
@@ -1,0 +1,1 @@
+export { MapControlId } from "./map-control-ids.enum.js";