+++ apps/frontend/src/pages/admin-dashboard/components/dashboard-layout/dashboard-layout.tsx
@@ -1,14 +1,14 @@
 import { type JSX, type PropsWithChildren } from "react";
 
 import { Sidebar } from "~/libs/components/components.js";
-import { NAVIGATION_ITEMS } from "~/libs/constants/constants.js";
+import { NAVIGATION_ITEMS_GROUPS } from "~/libs/constants/constants.js";
 
 import styles from "./styles.module.css";
 
 const DashboardLayout = ({ children }: PropsWithChildren): JSX.Element => {
 	return (
 		<div className={styles["container"]}>
-			<Sidebar navigationItems={NAVIGATION_ITEMS} />
+			<Sidebar navigationItemsGroups={NAVIGATION_ITEMS_GROUPS} />
 			<main className={styles["dashboard"]}>{children}</main>
 		</div>
 	);